Rename useVoiceRecorder hook to VoiceRecorder
